# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BaseBook DEX API
  description: |
    REST API for the BaseBook decentralized exchange on Base (Chain ID 8453).
    Provides swap quotes, pool/token data, liquidity positions, price oracles,
    chart data, and protocol analytics.

    **Authentication:** Most endpoints are public. Endpoints that require
    authentication expect an `x-api-key` header. The swap quote endpoint is
    rate-limited to 10 requests per minute per IP.
  version: 1.0.0
  contact:
    name: BaseBook Team
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.basebook.xyz
    description: Production
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authenticated endpoints
  schemas:
...
    EthAddress:
      type: string
      pattern: '^0x[a-fA-F0-9]{40}$'
      example: '0x4200000000000000000000000000000000000006'
    PoolId:
      type: string
      pattern: '^0x[a-fA-F0-9]{64}$'
      example: '0xabc123...64-hex-chars'
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

    # ── Health ──────────────────────────────────────────────────────
...
    PoolToken:
      type: object
      required: [id, address, symbol, name, decimals]
      properties:
        id:
          type: integer
        address:
          $ref: '#/components/schemas/EthAddress'
        symbol:
          type: string
        name:
          type: string
        decimals:
          type: integer
        logoUri:
          type: string
          nullable: true
    Pool:
      type: object
      required:
        - id
        - poolId
        - chainId
        - token0
        - token1
        - feeTier
        - tickSpacing
        - sqrtPriceX96
        - currentTick
        - liquidity
        - volume24hUsd
        - volume24hToken0
        - volume24hToken1
        - fees24hUsd
        - txCount24h
        - tvlUsd
        - tvlToken0
        - tvlToken1
        - apr24h
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        poolId:
          $ref: '#/components/schemas/PoolId'
        chainId:
          type: integer
        token0:
          $ref: '#/components/schemas/PoolToken'
        token1:
          $ref: '#/components/schemas/PoolToken'
        feeTier:
          type: integer
        tickSpacing:
          type: integer
        hookAddress:
          type: string
          nullable: true
        sqrtPriceX96:
          type: string
        currentTick:
          type: integer
        liquidity:
          type: string
        token0Price:
          type: string
          nullable: true
        token1Price:
          type: string
          nullable: true
        volume24hUsd:
          type: string
        volume24hToken0:
          type: string
        volume24hToken1:
          type: string
        fees24hUsd:
          type: string
        txCount24h:
          type: integer
        tvlUsd:
          type: string
        tvlToken0:
          type: string
        tvlToken1:
          type: string
        apr24h:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PoolDetail:
      allOf:
        - $ref: '#/components/schemas/Pool'
        - type: object
          properties:
            volume7dUsd:
              type: string
            volume30dUsd:
              type: string
            fees7dUsd:
              type: string
            fees30dUsd:
              type: string
    PoolListResponse:
      type: object
      required: [pools, total, limit, offset]
      properties:
        pools:
          type: array
          items:
            $ref: '#/components/schemas/Pool'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ── Positions ───────────────────────────────────────────────────
...
  parameters:
    ChainIdQuery:
      name: chainId
      in: query
      schema:
        type: integer
        default: 8453
      description: Chain ID (defaults to Base mainnet)
    PoolIdPath:
      name: poolId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/PoolId'
      description: Pool ID (keccak256 of PoolKey)
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

# Default security: API key required. Public endpoints override with []
security:
  - ApiKeyAuth: []
paths:
...
  /v1/pools:
    get:
      operationId: getPools
      tags: [Pools]
      summary: List liquidity pools
      description: Returns a paginated, sortable list of pools with optional token and TVL filters.
      security: []
      parameters:
        - $ref: '#/components/parameters/ChainIdQuery'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [tvl, volume, apr, createdAt]
            default: tvl
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: token0
          in: query
          schema:
            $ref: '#/components/schemas/EthAddress'
          description: Filter by token0 address
        - name: token1
          in: query
          schema:
            $ref: '#/components/schemas/EthAddress'
          description: Filter by token1 address
        - name: minTvl
          in: query
          schema:
            type: number
            minimum: 0
          description: Minimum TVL in USD
      responses:
        '200':
          description: Paginated pool list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/pools/{poolId}:
    get:
      operationId: getPool
      tags: [Pools]
      summary: Get pool details by ID
      security: []
      parameters:
        - $ref: '#/components/parameters/PoolIdPath'
        - $ref: '#/components/parameters/ChainIdQuery'
      responses:
        '200':
          description: Pool detail including extended stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ════════════════════════════════════════════════════════════════════
  # Positions
  # ════════════════════════════════════════════════════════════════════
...